tests/submission/PhilipTaylor/canvas/2d.getcontext.extraargs.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: 2d.getcontext.extraargs</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.2d.html">2d</a>.<a href="index.2d.getcontext.html">getcontext</a>.extraargs</h1>
<p class="desc">The 2D context ignores extra getContext arguments</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.context.2d.extraargs">context.2d.extraargs</a>
</li></ul>
</div>

<p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("The 2D context ignores extra getContext arguments");
_addTest(function(canvas, ctx) {

_assertDifferent(canvas.getContext('2d', false, {}, [], 1, "2"), null, "canvas.getContext('2d', false, {}, [], 1, \"2\")", "null");


});
</script>

<div id="log"></div>
</body></html>
